# simpeg-go/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.Cuti:
    properties:
      alasan:
        type: string
      catatan_persetujuan:
        type: string
      created_at:
        type: string
      disetujui_oleh:
        type: integer
      disetujui_pada:
        type: string
      id:
        type: integer
      jenis:
        type: string
      jumlah_hari:
        type: integer
      pegawai:
        $ref: '#/definitions/domain.Pegawai'
      pegawai_id:
        type: integer
      status:
        type: string
      tanggal_mulai:
        type: string
      tanggal_selesai:
        type: string
      updated_at:
        type: string
    type: object
  domain.DashboardStats:
    properties:
      absensi_hari_ini:
        type: integer
      cuti_menunggu:
        type: integer
      pegawai_aktif:
        type: integer
      total_jabatan:
        type: integer
      total_pegawai:
        type: integer
      total_unit_kerja:
        type: integer
    type: object
...
  domain.Golongan:
    properties:
      created_at:
        type: string
      id:
        type: integer
      kode:
        type: string
      nama:
        type: string
      updated_at:
        type: string
      urutan:
        type: integer
    type: object
  domain.Jabatan:
    properties:
      created_at:
        type: string
      id:
        type: integer
      is_aktif:
        type: boolean
      jenis:
        type: string
      kode:
        type: string
      nama:
        type: string
      unit_kerja_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  domain.Pegawai:
    properties:
      agama:
        type: string
      alamat_domisili:
        type: string
      alamat_ktp:
        type: string
      created_at:
        type: string
      created_by:
        type: integer
      email_institusi:
        type: string
      email_pribadi:
        type: string
      foto:
        type: string
      foto_url:
        type: string
      golongan:
        $ref: '#/definitions/domain.Golongan'
      golongan_id:
        type: integer
      hubungan_kontak_darurat:
        type: string
      id:
        type: integer
      jabatan:
        $ref: '#/definitions/domain.Jabatan'
      jabatan_id:
        type: integer
      jenis_kelamin:
        type: string
      kewarganegaraan:
        type: string
      nama_kontak_darurat:
        type: string
      nama_lengkap:
        type: string
      nama_panggilan:
        type: string
      nidn:
        type: string
      nik:
        type: string
      nip:
        type: string
      nomor_bpjs_kesehatan:
        type: string
      nomor_bpjs_ketenagakerjaan:
        type: string
      nomor_hp:
        type: string
      nomor_telp_darurat:
        type: string
      npwp:
        type: string
      status_aktif:
        type: string
      status_kepegawaian:
        allOf:
        - $ref: '#/definitions/domain.StatusKepegawaian'
        description: Relations (for detail view)
      status_kepegawaian_id:
        type: integer
      status_pernikahan:
        type: string
      tanggal_lahir:
        type: string
      tanggal_masuk:
        type: string
      tanggal_pensiun:
        type: string
      tanggal_tmt_pns:
        type: string
      tempat_lahir:
        type: string
      unit_kerja:
        $ref: '#/definitions/domain.UnitKerja'
      unit_kerja_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: integer
      user_id:
        type: integer
    type: object
...
  domain.StatusKepegawaian:
    properties:
      created_at:
        type: string
      id:
        type: integer
      is_aktif:
        type: boolean
      kode:
        type: string
      nama:
        type: string
      updated_at:
        type: string
    type: object
  domain.UnitKerja:
    properties:
      children:
        items:
          $ref: '#/definitions/domain.UnitKerja'
        type: array
      created_at:
        type: string
      id:
        type: integer
      is_aktif:
        type: boolean
      kode:
        type: string
      nama:
        type: string
      parent_id:
        type: integer
      singkatan:
        type: string
      tipe:
        type: string
      updated_at:
        type: string
    type: object
...
  http.CutiApproveRequest:
    properties:
      catatan:
        type: string
      status:
        type: string
    type: object
  http.CutiStoreRequest:
    properties:
      alasan:
        type: string
      jenis:
        type: string
      jumlah_hari:
        type: integer
      pegawai_id:
        type: integer
      tanggal_mulai:
        type: string
      tanggal_selesai:
        type: string
    type: object
...
  http.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  http.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    name: SIMPEG UML
  description: Backend API Sistem Informasi Manajemen Kepegawaian Universitas Muhammadiyah
    Lampung (Gin + GORM).
  termsOfService: http://simpeg.uml.ac.id
  title: SIMPEG UML API
  version: "1.0"
paths:
...
  /cuti:
    get:
      description: Tanpa `pegawai_id` mengembalikan list semua cuti (paginasi, filter
        status). Dengan `pegawai_id` mengembalikan cuti milik pegawai tersebut.
      parameters:
      - description: Filter ID pegawai
        in: query
        name: pegawai_id
        type: integer
      - description: Halaman (default 1)
        in: query
        name: page
        type: integer
      - description: Jumlah per halaman (default 20, maks 100)
        in: query
        name: limit
        type: integer
      - description: Filter status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Cuti'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Daftar cuti
      tags:
      - Cuti
    post:
      consumes:
      - application/json
      description: 'Self-service: admin/operator (HRD) bisa ajukan atas nama siapa
        pun; pegawai/dosen/pimpinan WAJIB atas nama sendiri (pegawai_id dipaksa milik
        akun yang login). Jenis cuti: tahunan, sakit, melahirkan, besar, alasan_penting.'
      parameters:
      - description: Data pengajuan cuti
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.CutiStoreRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Cuti'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Ajukan cuti
      tags:
      - Cuti
  /cuti/{id}:
    patch:
      consumes:
      - application/json
      description: 'Status: `disetujui` atau `ditolak`.'
      parameters:
      - description: ID cuti
        in: path
        name: id
        required: true
        type: integer
      - description: Keputusan verifikasi
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.CutiApproveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Setujui/tolak cuti
      tags:
      - Cuti
  /dashboard/stats:
    get:
      description: Ringkasan jumlah pegawai, unit kerja, golongan, dan metrik lainnya.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.DashboardStats'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Statistik dashboard
      tags:
      - Dashboard
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"